cdk/lambda/dynamodb-ope-queue/requirements.txt
ulid_py==1.1.0
